restructuring folder & installing packages
