+++ requirements.txt
@@ -4,3 +4,5 @@ Flask==3.0.3
 gunicorn==22.0.0
 Werkzeug==3.0.6
 Flask-JWT-Extended==4.7.1
+Flask-SQLAlchemy==3.1.1
+PyMySQL==1.1.1
